fix: unable to factory reset when nitro is running on windows
