+++ web/helpers/atoms/App.atom.ts
@@ -3,3 +3,5 @@ import { atom } from 'jotai'
 import { MainViewState } from '@/constants/screens'
 
 export const mainViewStateAtom = atom<MainViewState>(MainViewState.Thread)
+
+export const defaultJanDataFolderAtom = atom<string>('')
